+++ esphome/components/api/api_pb2.h
@@ -4,6 +4,9 @@
 
 #include "proto.h"
 #include "api_pb2_size.h"
+#ifdef HAS_PROTO_MESSAGE_DUMP
+#include "api_pb2_dump.h"
+#endif
 
 namespace esphome {
 namespace api {
@@ -15,6 +18,7 @@ enum EntityCategory : uint32_t {
   ENTITY_CATEGORY_CONFIG = 1,
   ENTITY_CATEGORY_DIAGNOSTIC = 2,
 };
+#ifdef USE_COVER
 enum LegacyCoverState : uint32_t {
   LEGACY_COVER_STATE_OPEN = 0,
   LEGACY_COVER_STATE_CLOSED = 1,
@@ -29,6 +33,8 @@ enum LegacyCoverCommand : uint32_t {
   LEGACY_COVER_COMMAND_CLOSE = 1,
   LEGACY_COVER_COMMAND_STOP = 2,
 };
+#endif
+#ifdef USE_FAN
 enum FanSpeed : uint32_t {
   FAN_SPEED_LOW = 0,
   FAN_SPEED_MEDIUM = 1,
@@ -38,6 +44,8 @@ enum FanDirection : uint32_t {
   FAN_DIRECTION_FORWARD = 0,
   FAN_DIRECTION_REVERSE = 1,
 };
+#endif
+#ifdef USE_LIGHT
 enum ColorMode : uint32_t {
   COLOR_MODE_UNKNOWN = 0,
   COLOR_MODE_ON_OFF = 1,
@@ -51,6 +59,8 @@ enum ColorMode : uint32_t {
   COLOR_MODE_RGB_COLOR_TEMPERATURE = 47,
   COLOR_MODE_RGB_COLD_WARM_WHITE = 51,
 };
+#endif
+#ifdef USE_SENSOR
 enum SensorStateClass : uint32_t {
   STATE_CLASS_NONE = 0,
   STATE_CLASS_MEASUREMENT = 1,
@@ -62,6 +72,7 @@ enum SensorLastResetType : uint32_t {
   LAST_RESET_NEVER = 1,
   LAST_RESET_AUTO = 2,
 };
+#endif
 enum LogLevel : uint32_t {
   LOG_LEVEL_NONE = 0,
   LOG_LEVEL_ERROR = 1,
@@ -82,6 +93,7 @@ enum ServiceArgType : uint32_t {
   SERVICE_ARG_TYPE_FLOAT_ARRAY = 6,
   SERVICE_ARG_TYPE_STRING_ARRAY = 7,
 };
+#ifdef USE_CLIMATE
 enum ClimateMode : uint32_t {
   CLIMATE_MODE_OFF = 0,
   CLIMATE_MODE_HEAT_COOL = 1,
@@ -127,11 +139,15 @@ enum ClimatePreset : uint32_t {
   CLIMATE_PRESET_SLEEP = 6,
   CLIMATE_PRESET_ACTIVITY = 7,
 };
+#endif
+#ifdef USE_NUMBER
 enum NumberMode : uint32_t {
   NUMBER_MODE_AUTO = 0,
   NUMBER_MODE_BOX = 1,
   NUMBER_MODE_SLIDER = 2,
 };
+#endif
+#ifdef USE_LOCK
 enum LockState : uint32_t {
   LOCK_STATE_NONE = 0,
   LOCK_STATE_LOCKED = 1,
@@ -145,6 +161,8 @@ enum LockCommand : uint32_t {
   LOCK_LOCK = 1,
   LOCK_OPEN = 2,
 };
+#endif
+#ifdef USE_MEDIA_PLAYER
 enum MediaPlayerState : uint32_t {
   MEDIA_PLAYER_STATE_NONE = 0,
   MEDIA_PLAYER_STATE_IDLE = 1,
@@ -162,6 +180,8 @@ enum MediaPlayerFormatPurpose : uint32_t {
   MEDIA_PLAYER_FORMAT_PURPOSE_DEFAULT = 0,
   MEDIA_PLAYER_FORMAT_PURPOSE_ANNOUNCEMENT = 1,
 };
+#endif
+#ifdef USE_BLUETOOTH_PROXY
 enum BluetoothDeviceRequestType : uint32_t {
   BLUETOOTH_DEVICE_REQUEST_TYPE_CONNECT = 0,
   BLUETOOTH_DEVICE_REQUEST_TYPE_DISCONNECT = 1,
@@ -183,6 +203,7 @@ enum BluetoothScannerMode : uint32_t {
   BLUETOOTH_SCANNER_MODE_PASSIVE = 0,
   BLUETOOTH_SCANNER_MODE_ACTIVE = 1,
 };
+#endif
 enum VoiceAssistantSubscribeFlag : uint32_t {
   VOICE_ASSISTANT_SUBSCRIBE_NONE = 0,
   VOICE_ASSISTANT_SUBSCRIBE_API_AUDIO = 1,
@@ -192,6 +213,7 @@ enum VoiceAssistantRequestFlag : uint32_t {
   VOICE_ASSISTANT_REQUEST_USE_VAD = 1,
   VOICE_ASSISTANT_REQUEST_USE_WAKE_WORD = 2,
 };
+#ifdef USE_VOICE_ASSISTANT
 enum VoiceAssistantEvent : uint32_t {
   VOICE_ASSISTANT_ERROR = 0,
   VOICE_ASSISTANT_RUN_START = 1,
@@ -216,6 +238,8 @@ enum VoiceAssistantTimerEvent : uint32_t {
   VOICE_ASSISTANT_TIMER_CANCELLED = 2,
   VOICE_ASSISTANT_TIMER_FINISHED = 3,
 };
+#endif
+#ifdef USE_ALARM_CONTROL_PANEL
 enum AlarmControlPanelState : uint32_t {
   ALARM_STATE_DISARMED = 0,
   ALARM_STATE_ARMED_HOME = 1,
@@ -237,20 +261,27 @@ enum AlarmControlPanelStateCommand : uint32_t {
   ALARM_CONTROL_PANEL_ARM_CUSTOM_BYPASS = 5,
   ALARM_CONTROL_PANEL_TRIGGER = 6,
 };
+#endif
+#ifdef USE_TEXT
 enum TextMode : uint32_t {
   TEXT_MODE_TEXT = 0,
   TEXT_MODE_PASSWORD = 1,
 };
+#endif
+#ifdef USE_VALVE
 enum ValveOperation : uint32_t {
   VALVE_OPERATION_IDLE = 0,
   VALVE_OPERATION_IS_OPENING = 1,
   VALVE_OPERATION_IS_CLOSING = 2,
 };
+#endif
+#ifdef USE_UPDATE
 enum UpdateCommand : uint32_t {
   UPDATE_COMMAND_NONE = 0,
   UPDATE_COMMAND_UPDATE = 1,
   UPDATE_COMMAND_CHECK = 2,
 };
+#endif
 
 }  // namespace enums
 
@@ -281,7 +312,7 @@ class HelloRequest : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 1;
   static constexpr uint16_t ESTIMATED_SIZE = 17;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "hello_request"; }
+  const char *message_name() const override;
 #endif
   std::string client_info{};
   uint32_t api_version_major{0};
@@ -300,7 +331,7 @@ class HelloResponse : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 2;
   static constexpr uint16_t ESTIMATED_SIZE = 26;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "hello_response"; }
+  const char *message_name() const override;
 #endif
   uint32_t api_version_major{0};
   uint32_t api_version_minor{0};
@@ -320,7 +351,7 @@ class ConnectRequest : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 3;
   static constexpr uint16_t ESTIMATED_SIZE = 9;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "connect_request"; }
+  const char *message_name() const override;
 #endif
   std::string password{};
   void encode(ProtoWriteBuffer buffer) const override;
@@ -337,7 +368,7 @@ class ConnectResponse : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 4;
   static constexpr uint16_t ESTIMATED_SIZE = 2;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "connect_response"; }
+  const char *message_name() const override;
 #endif
   bool invalid_password{false};
   void encode(ProtoWriteBuffer buffer) const override;
@@ -354,7 +385,7 @@ class DisconnectRequest : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 5;
   static constexpr uint16_t ESTIMATED_SIZE = 0;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "disconnect_request"; }
+  const char *message_name() const override;
 #endif
 #ifdef HAS_PROTO_MESSAGE_DUMP
   void dump_to(std::string &out) const override;
@@ -367,7 +398,7 @@ class DisconnectResponse : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 6;
   static constexpr uint16_t ESTIMATED_SIZE = 0;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "disconnect_response"; }
+  const char *message_name() const override;
 #endif
 #ifdef HAS_PROTO_MESSAGE_DUMP
   void dump_to(std::string &out) const override;
@@ -380,7 +411,7 @@ class PingRequest : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 7;
   static constexpr uint16_t ESTIMATED_SIZE = 0;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "ping_request"; }
+  const char *message_name() const override;
 #endif
 #ifdef HAS_PROTO_MESSAGE_DUMP
   void dump_to(std::string &out) const override;
@@ -393,7 +424,7 @@ class PingResponse : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 8;
   static constexpr uint16_t ESTIMATED_SIZE = 0;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "ping_response"; }
+  const char *message_name() const override;
 #endif
 #ifdef HAS_PROTO_MESSAGE_DUMP
   void dump_to(std::string &out) const override;
@@ -406,7 +437,7 @@ class DeviceInfoRequest : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 9;
   static constexpr uint16_t ESTIMATED_SIZE = 0;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "device_info_request"; }
+  const char *message_name() const override;
 #endif
 #ifdef HAS_PROTO_MESSAGE_DUMP
   void dump_to(std::string &out) const override;
@@ -446,7 +477,7 @@ class DeviceInfoResponse : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 10;
   static constexpr uint16_t ESTIMATED_SIZE = 219;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "device_info_response"; }
+  const char *message_name() const override;
 #endif
   bool uses_password{false};
   std::string name{};
@@ -484,7 +515,7 @@ class ListEntitiesRequest : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 11;
   static constexpr uint16_t ESTIMATED_SIZE = 0;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "list_entities_request"; }
+  const char *message_name() const override;
 #endif
 #ifdef HAS_PROTO_MESSAGE_DUMP
   void dump_to(std::string &out) const override;
@@ -497,7 +528,7 @@ class ListEntitiesDoneResponse : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 19;
   static constexpr uint16_t ESTIMATED_SIZE = 0;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "list_entities_done_response"; }
+  const char *message_name() const override;
 #endif
 #ifdef HAS_PROTO_MESSAGE_DUMP
   void dump_to(std::string &out) const override;
@@ -510,7 +541,7 @@ class SubscribeStatesRequest : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 20;
   static constexpr uint16_t ESTIMATED_SIZE = 0;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "subscribe_states_request"; }
+  const char *message_name() const override;
 #endif
 #ifdef HAS_PROTO_MESSAGE_DUMP
   void dump_to(std::string &out) const override;
@@ -518,12 +549,13 @@ class SubscribeStatesRequest : public ProtoMessage {
 
  protected:
 };
+#ifdef USE_BINARY_SENSOR
 class ListEntitiesBinarySensorResponse : public InfoResponseProtoMessage {
  public:
   static constexpr uint16_t MESSAGE_TYPE = 12;
   static constexpr uint16_t ESTIMATED_SIZE = 60;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "list_entities_binary_sensor_response"; }
+  const char *message_name() const override;
 #endif
   std::string device_class{};
   bool is_status_binary_sensor{false};
@@ -541,7 +573,7 @@ class BinarySensorStateResponse : public StateResponseProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 21;
   static constexpr uint16_t ESTIMATED_SIZE = 9;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "binary_sensor_state_response"; }
+  const char *message_name() const override;
 #endif
   bool state{false};
   bool missing_state{false};
@@ -554,12 +586,14 @@ class BinarySensorStateResponse : public StateResponseProtoMessage {
  protected:
   bool decode_field(uint32_t field_id, ProtoFieldValue value) override;
 };
+#endif
+#ifdef USE_COVER
 class ListEntitiesCoverResponse : public InfoResponseProtoMessage {
  public:
   static constexpr uint16_t MESSAGE_TYPE = 13;
   static constexpr uint16_t ESTIMATED_SIZE = 66;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "list_entities_cover_response"; }
+  const char *message_name() const override;
 #endif
   bool assumed_state{false};
   bool supports_position{false};
@@ -580,7 +614,7 @@ class CoverStateResponse : public StateResponseProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 22;
   static constexpr uint16_t ESTIMATED_SIZE = 19;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "cover_state_response"; }
+  const char *message_name() const override;
 #endif
   enums::LegacyCoverState legacy_state{};
   float position{0.0f};
@@ -600,7 +634,7 @@ class CoverCommandRequest : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 30;
   static constexpr uint16_t ESTIMATED_SIZE = 25;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "cover_command_request"; }
+  const char *message_name() const override;
 #endif
   uint32_t key{0};
   bool has_legacy_command{false};
@@ -619,12 +653,14 @@ class CoverCommandRequest : public ProtoMessage {
  protected:
   bool decode_field(uint32_t field_id, ProtoFieldValue value) override;
 };
+#endif
+#ifdef USE_FAN
 class ListEntitiesFanResponse : public InfoResponseProtoMessage {
  public:
   static constexpr uint16_t MESSAGE_TYPE = 14;
   static constexpr uint16_t ESTIMATED_SIZE = 77;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "list_entities_fan_response"; }
+  const char *message_name() const override;
 #endif
   bool supports_oscillation{false};
   bool supports_speed{false};
@@ -645,7 +681,7 @@ class FanStateResponse : public StateResponseProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 23;
   static constexpr uint16_t ESTIMATED_SIZE = 26;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "fan_state_response"; }
+  const char *message_name() const override;
 #endif
   bool state{false};
   bool oscillating{false};
@@ -667,7 +703,7 @@ class FanCommandRequest : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 31;
   static constexpr uint16_t ESTIMATED_SIZE = 38;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "fan_command_request"; }
+  const char *message_name() const override;
 #endif
   uint32_t key{0};
   bool has_state{false};
@@ -691,12 +727,14 @@ class FanCommandRequest : public ProtoMessage {
  protected:
   bool decode_field(uint32_t field_id, ProtoFieldValue value) override;
 };
+#endif
+#ifdef USE_LIGHT
 class ListEntitiesLightResponse : public InfoResponseProtoMessage {
  public:
   static constexpr uint16_t MESSAGE_TYPE = 15;
   static constexpr uint16_t ESTIMATED_SIZE = 90;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "list_entities_light_response"; }
+  const char *message_name() const override;
 #endif
   std::vector<enums::ColorMode> supported_color_modes{};
   bool legacy_supports_brightness{false};
@@ -720,7 +758,7 @@ class LightStateResponse : public StateResponseProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 24;
   static constexpr uint16_t ESTIMATED_SIZE = 63;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "light_state_response"; }
+  const char *message_name() const override;
 #endif
   bool state{false};
   float brightness{0.0f};
@@ -748,7 +786,7 @@ class LightCommandRequest : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 32;
   static constexpr uint16_t ESTIMATED_SIZE = 107;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "light_command_request"; }
+  const char *message_name() const override;
 #endif
   uint32_t key{0};
   bool has_state{false};
@@ -786,12 +824,14 @@ class LightCommandRequest : public ProtoMessage {
  protected:
   bool decode_field(uint32_t field_id, ProtoFieldValue value) override;
 };
+#endif
+#ifdef USE_SENSOR
 class ListEntitiesSensorResponse : public InfoResponseProtoMessage {
  public:
   static constexpr uint16_t MESSAGE_TYPE = 16;
   static constexpr uint16_t ESTIMATED_SIZE = 77;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "list_entities_sensor_response"; }
+  const char *message_name() const override;
 #endif
   std::string unit_of_measurement{};
   int32_t accuracy_decimals{0};
@@ -813,7 +853,7 @@ class SensorStateResponse : public StateResponseProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 25;
   static constexpr uint16_t ESTIMATED_SIZE = 12;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "sensor_state_response"; }
+  const char *message_name() const override;
 #endif
   float state{0.0f};
   bool missing_state{false};
@@ -826,12 +866,14 @@ class SensorStateResponse : public StateResponseProtoMessage {
  protected:
   bool decode_field(uint32_t field_id, ProtoFieldValue value) override;
 };
+#endif
+#ifdef USE_SWITCH
 class ListEntitiesSwitchResponse : public InfoResponseProtoMessage {
  public:
   static constexpr uint16_t MESSAGE_TYPE = 17;
   static constexpr uint16_t ESTIMATED_SIZE = 60;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "list_entities_switch_response"; }
+  const char *message_name() const override;
 #endif
   bool assumed_state{false};
   std::string device_class{};
@@ -849,7 +891,7 @@ class SwitchStateResponse : public StateResponseProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 26;
   static constexpr uint16_t ESTIMATED_SIZE = 7;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "switch_state_response"; }
+  const char *message_name() const override;
 #endif
   bool state{false};
   void encode(ProtoWriteBuffer buffer) const override;
@@ -866,7 +908,7 @@ class SwitchCommandRequest : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 33;
   static constexpr uint16_t ESTIMATED_SIZE = 7;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "switch_command_request"; }
+  const char *message_name() const override;
 #endif
   uint32_t key{0};
   bool state{false};
@@ -879,12 +921,14 @@ class SwitchCommandRequest : public ProtoMessage {
  protected:
   bool decode_field(uint32_t field_id, ProtoFieldValue value) override;
 };
+#endif
+#ifdef USE_TEXT_SENSOR
 class ListEntitiesTextSensorResponse : public InfoResponseProtoMessage {
  public:
   static constexpr uint16_t MESSAGE_TYPE = 18;
   static constexpr uint16_t ESTIMATED_SIZE = 58;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "list_entities_text_sensor_response"; }
+  const char *message_name() const override;
 #endif
   std::string device_class{};
   void encode(ProtoWriteBuffer buffer) const override;
@@ -901,7 +945,7 @@ class TextSensorStateResponse : public StateResponseProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 27;
   static constexpr uint16_t ESTIMATED_SIZE = 16;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "text_sensor_state_response"; }
+  const char *message_name() const override;
 #endif
   std::string state{};
   bool missing_state{false};
@@ -914,12 +958,13 @@ class TextSensorStateResponse : public StateResponseProtoMessage {
  protected:
   bool decode_field(uint32_t field_id, ProtoFieldValue value) override;
 };
+#endif
 class SubscribeLogsRequest : public ProtoMessage {
  public:
   static constexpr uint16_t MESSAGE_TYPE = 28;
   static constexpr uint16_t ESTIMATED_SIZE = 4;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "subscribe_logs_request"; }
+  const char *message_name() const override;
 #endif
   enums::LogLevel level{};
   bool dump_config{false};
@@ -937,7 +982,7 @@ class SubscribeLogsResponse : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 29;
   static constexpr uint16_t ESTIMATED_SIZE = 13;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "subscribe_logs_response"; }
+  const char *message_name() const override;
 #endif
   enums::LogLevel level{};
   std::string message{};
@@ -951,12 +996,13 @@ class SubscribeLogsResponse : public ProtoMessage {
  protected:
   bool decode_field(uint32_t field_id, ProtoFieldValue value) override;
 };
+#ifdef USE_API_NOISE
 class NoiseEncryptionSetKeyRequest : public ProtoMessage {
  public:
   static constexpr uint16_t MESSAGE_TYPE = 124;
   static constexpr uint16_t ESTIMATED_SIZE = 9;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "noise_encryption_set_key_request"; }
+  const char *message_name() const override;
 #endif
   std::string key{};
   void encode(ProtoWriteBuffer buffer) const override;
@@ -973,7 +1019,7 @@ class NoiseEncryptionSetKeyResponse : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 125;
   static constexpr uint16_t ESTIMATED_SIZE = 2;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "noise_encryption_set_key_response"; }
+  const char *message_name() const override;
 #endif
   bool success{false};
   void encode(ProtoWriteBuffer buffer) const override;
@@ -985,12 +1031,13 @@ class NoiseEncryptionSetKeyResponse : public ProtoMessage {
  protected:
   bool decode_field(uint32_t field_id, ProtoFieldValue value) override;
 };
+#endif
 class SubscribeHomeassistantServicesRequest : public ProtoMessage {
  public:
   static constexpr uint16_t MESSAGE_TYPE = 34;
   static constexpr uint16_t ESTIMATED_SIZE = 0;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "subscribe_homeassistant_services_request"; }
+  const char *message_name() const override;
 #endif
 #ifdef HAS_PROTO_MESSAGE_DUMP
   void dump_to(std::string &out) const override;
@@ -1016,7 +1063,7 @@ class HomeassistantServiceResponse : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 35;
   static constexpr uint16_t ESTIMATED_SIZE = 113;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "homeassistant_service_response"; }
+  const char *message_name() const override;
 #endif
   std::string service{};
   std::vector<HomeassistantServiceMap> data{};
@@ -1037,7 +1084,7 @@ class SubscribeHomeAssistantStatesRequest : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 38;
   static constexpr uint16_t ESTIMATED_SIZE = 0;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "subscribe_home_assistant_states_request"; }
+  const char *message_name() const override;
 #endif
 #ifdef HAS_PROTO_MESSAGE_DUMP
   void dump_to(std::string &out) const override;
@@ -1050,7 +1097,7 @@ class SubscribeHomeAssistantStateResponse : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 39;
   static constexpr uint16_t ESTIMATED_SIZE = 20;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "subscribe_home_assistant_state_response"; }
+  const char *message_name() const override;
 #endif
   std::string entity_id{};
   std::string attribute{};
@@ -1069,7 +1116,7 @@ class HomeAssistantStateResponse : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 40;
   static constexpr uint16_t ESTIMATED_SIZE = 27;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "home_assistant_state_response"; }
+  const char *message_name() const override;
 #endif
   std::string entity_id{};
   std::string state{};
@@ -1088,7 +1135,7 @@ class GetTimeRequest : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 36;
   static constexpr uint16_t ESTIMATED_SIZE = 0;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "get_time_request"; }
+  const char *message_name() const override;
 #endif
 #ifdef HAS_PROTO_MESSAGE_DUMP
   void dump_to(std::string &out) const override;
@@ -1101,7 +1148,7 @@ class GetTimeResponse : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 37;
   static constexpr uint16_t ESTIMATED_SIZE = 5;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "get_time_response"; }
+  const char *message_name() const override;
 #endif
   uint32_t epoch_seconds{0};
   void encode(ProtoWriteBuffer buffer) const override;
@@ -1131,7 +1178,7 @@ class ListEntitiesServicesResponse : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 41;
   static constexpr uint16_t ESTIMATED_SIZE = 48;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "list_entities_services_response"; }
+  const char *message_name() const override;
 #endif
   std::string name{};
   uint32_t key{0};
@@ -1170,7 +1217,7 @@ class ExecuteServiceRequest : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 42;
   static constexpr uint16_t ESTIMATED_SIZE = 39;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "execute_service_request"; }
+  const char *message_name() const override;
 #endif
   uint32_t key{0};
   std::vector<ExecuteServiceArgument> args{};
@@ -1183,12 +1230,13 @@ class ExecuteServiceRequest : public ProtoMessage {
  protected:
   bool decode_field(uint32_t field_id, ProtoFieldValue value) override;
 };
+#ifdef USE_ESP32_CAMERA
 class ListEntitiesCameraResponse : public InfoResponseProtoMessage {
  public:
   static constexpr uint16_t MESSAGE_TYPE = 43;
   static constexpr uint16_t ESTIMATED_SIZE = 49;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "list_entities_camera_response"; }
+  const char *message_name() const override;
 #endif
   void encode(ProtoWriteBuffer buffer) const override;
   void calculate_size(uint32_t &total_size) const override;
@@ -1204,7 +1252,7 @@ class CameraImageResponse : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 44;
   static constexpr uint16_t ESTIMATED_SIZE = 16;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "camera_image_response"; }
+  const char *message_name() const override;
 #endif
   uint32_t key{0};
   std::string data{};
@@ -1223,7 +1271,7 @@ class CameraImageRequest : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 45;
   static constexpr uint16_t ESTIMATED_SIZE = 4;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "camera_image_request"; }
+  const char *message_name() const override;
 #endif
   bool single{false};
   bool stream{false};
@@ -1236,12 +1284,14 @@ class CameraImageRequest : public ProtoMessage {
  protected:
   bool decode_field(uint32_t field_id, ProtoFieldValue value) override;
 };
+#endif
+#ifdef USE_CLIMATE
 class ListEntitiesClimateResponse : public InfoResponseProtoMessage {
  public:
   static constexpr uint16_t MESSAGE_TYPE = 46;
   static constexpr uint16_t ESTIMATED_SIZE = 156;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "list_entities_climate_response"; }
+  const char *message_name() const override;
 #endif
   bool supports_current_temperature{false};
   bool supports_two_point_target_temperature{false};
@@ -1275,7 +1325,7 @@ class ClimateStateResponse : public StateResponseProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 47;
   static constexpr uint16_t ESTIMATED_SIZE = 65;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "climate_state_response"; }
+  const char *message_name() const override;
 #endif
   enums::ClimateMode mode{};
   float current_temperature{0.0f};
@@ -1305,7 +1355,7 @@ class ClimateCommandRequest : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 48;
   static constexpr uint16_t ESTIMATED_SIZE = 83;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "climate_command_request"; }
+  const char *message_name() const override;
 #endif
   uint32_t key{0};
   bool has_mode{false};
@@ -1339,12 +1389,14 @@ class ClimateCommandRequest : public ProtoMessage {
  protected:
   bool decode_field(uint32_t field_id, ProtoFieldValue value) override;
 };
+#endif
+#ifdef USE_NUMBER
 class ListEntitiesNumberResponse : public InfoResponseProtoMessage {
  public:
   static constexpr uint16_t MESSAGE_TYPE = 49;
   static constexpr uint16_t ESTIMATED_SIZE = 84;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "list_entities_number_response"; }
+  const char *message_name() const override;
 #endif
   float min_value{0.0f};
   float max_value{0.0f};
@@ -1366,7 +1418,7 @@ class NumberStateResponse : public StateResponseProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 50;
   static constexpr uint16_t ESTIMATED_SIZE = 12;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "number_state_response"; }
+  const char *message_name() const override;
 #endif
   float state{0.0f};
   bool missing_state{false};
@@ -1384,7 +1436,7 @@ class NumberCommandRequest : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 51;
   static constexpr uint16_t ESTIMATED_SIZE = 10;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "number_command_request"; }
+  const char *message_name() const override;
 #endif
   uint32_t key{0};
   float state{0.0f};
@@ -1397,12 +1449,14 @@ class NumberCommandRequest : public ProtoMessage {
  protected:
   bool decode_field(uint32_t field_id, ProtoFieldValue value) override;
 };
+#endif
+#ifdef USE_SELECT
 class ListEntitiesSelectResponse : public InfoResponseProtoMessage {
  public:
   static constexpr uint16_t MESSAGE_TYPE = 52;
   static constexpr uint16_t ESTIMATED_SIZE = 67;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "list_entities_select_response"; }
+  const char *message_name() const override;
 #endif
   std::vector<std::string> options{};
   void encode(ProtoWriteBuffer buffer) const override;
@@ -1419,7 +1473,7 @@ class SelectStateResponse : public StateResponseProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 53;
   static constexpr uint16_t ESTIMATED_SIZE = 16;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "select_state_response"; }
+  const char *message_name() const override;
 #endif
   std::string state{};
   bool missing_state{false};
@@ -1437,7 +1491,7 @@ class SelectCommandRequest : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 54;
   static constexpr uint16_t ESTIMATED_SIZE = 14;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "select_command_request"; }
+  const char *message_name() const override;
 #endif
   uint32_t key{0};
   std::string state{};
@@ -1450,12 +1504,14 @@ class SelectCommandRequest : public ProtoMessage {
  protected:
   bool decode_field(uint32_t field_id, ProtoFieldValue value) override;
 };
+#endif
+#ifdef USE_SIREN
 class ListEntitiesSirenResponse : public InfoResponseProtoMessage {
  public:
   static constexpr uint16_t MESSAGE_TYPE = 55;
   static constexpr uint16_t ESTIMATED_SIZE = 71;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "list_entities_siren_response"; }
+  const char *message_name() const override;
 #endif
   std::vector<std::string> tones{};
   bool supports_duration{false};
@@ -1474,7 +1530,7 @@ class SirenStateResponse : public StateResponseProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 56;
   static constexpr uint16_t ESTIMATED_SIZE = 7;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "siren_state_response"; }
+  const char *message_name() const override;
 #endif
   bool state{false};
   void encode(ProtoWriteBuffer buffer) const override;
@@ -1491,7 +1547,7 @@ class SirenCommandRequest : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 57;
   static constexpr uint16_t ESTIMATED_SIZE = 33;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "siren_command_request"; }
+  const char *message_name() const override;
 #endif
   uint32_t key{0};
   bool has_state{false};
@@ -1511,12 +1567,14 @@ class SirenCommandRequest : public ProtoMessage {
  protected:
   bool decode_field(uint32_t field_id, ProtoFieldValue value) override;
 };
+#endif
+#ifdef USE_LOCK
 class ListEntitiesLockResponse : public InfoResponseProtoMessage {
  public:
   static constexpr uint16_t MESSAGE_TYPE = 58;
   static constexpr uint16_t ESTIMATED_SIZE = 64;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "list_entities_lock_response"; }
+  const char *message_name() const override;
 #endif
   bool assumed_state{false};
   bool supports_open{false};
@@ -1536,7 +1594,7 @@ class LockStateResponse : public StateResponseProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 59;
   static constexpr uint16_t ESTIMATED_SIZE = 7;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "lock_state_response"; }
+  const char *message_name() const override;
 #endif
   enums::LockState state{};
   void encode(ProtoWriteBuffer buffer) const override;
@@ -1553,7 +1611,7 @@ class LockCommandRequest : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 60;
   static constexpr uint16_t ESTIMATED_SIZE = 18;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "lock_command_request"; }
+  const char *message_name() const override;
 #endif
   uint32_t key{0};
   enums::LockCommand command{};
@@ -1568,12 +1626,14 @@ class LockCommandRequest : public ProtoMessage {
  protected:
   bool decode_field(uint32_t field_id, ProtoFieldValue value) override;
 };
+#endif
+#ifdef USE_BUTTON
 class ListEntitiesButtonResponse : public InfoResponseProtoMessage {
  public:
   static constexpr uint16_t MESSAGE_TYPE = 61;
   static constexpr uint16_t ESTIMATED_SIZE = 58;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "list_entities_button_response"; }
+  const char *message_name() const override;
 #endif
   std::string device_class{};
   void encode(ProtoWriteBuffer buffer) const override;
@@ -1590,7 +1650,7 @@ class ButtonCommandRequest : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 62;
   static constexpr uint16_t ESTIMATED_SIZE = 5;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "button_command_request"; }
+  const char *message_name() const override;
 #endif
   uint32_t key{0};
   void encode(ProtoWriteBuffer buffer) const override;
@@ -1602,6 +1662,8 @@ class ButtonCommandRequest : public ProtoMessage {
  protected:
   bool decode_field(uint32_t field_id, ProtoFieldValue value) override;
 };
+#endif
+#ifdef USE_MEDIA_PLAYER
 class MediaPlayerSupportedFormat : public ProtoMessage {
  public:
   std::string format{};
@@ -1623,7 +1685,7 @@ class ListEntitiesMediaPlayerResponse : public InfoResponseProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 63;
   static constexpr uint16_t ESTIMATED_SIZE = 85;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "list_entities_media_player_response"; }
+  const char *message_name() const override;
 #endif
   bool supports_pause{false};
   std::vector<MediaPlayerSupportedFormat> supported_formats{};
@@ -1641,7 +1703,7 @@ class MediaPlayerStateResponse : public StateResponseProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 64;
   static constexpr uint16_t ESTIMATED_SIZE = 14;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "media_player_state_response"; }
+  const char *message_name() const override;
 #endif
   enums::MediaPlayerState state{};
   float volume{0.0f};
@@ -1660,7 +1722,7 @@ class MediaPlayerCommandRequest : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 65;
   static constexpr uint16_t ESTIMATED_SIZE = 31;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "media_player_command_request"; }
+  const char *message_name() const override;
 #endif
   uint32_t key{0};
   bool has_command{false};
@@ -1680,12 +1742,14 @@ class MediaPlayerCommandRequest : public ProtoMessage {
  protected:
   bool decode_field(uint32_t field_id, ProtoFieldValue value) override;
 };
+#endif
+#ifdef USE_BLUETOOTH_PROXY
 class SubscribeBluetoothLEAdvertisementsRequest : public ProtoMessage {
  public:
   static constexpr uint16_t MESSAGE_TYPE = 66;
   static constexpr uint16_t ESTIMATED_SIZE = 4;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "subscribe_bluetooth_le_advertisements_request"; }
+  const char *message_name() const override;
 #endif
   uint32_t flags{0};
   void encode(ProtoWriteBuffer buffer) const override;
@@ -1716,7 +1780,7 @@ class BluetoothLEAdvertisementResponse : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 67;
   static constexpr uint16_t ESTIMATED_SIZE = 107;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "bluetooth_le_advertisement_response"; }
+  const char *message_name() const override;
 #endif
   uint64_t address{0};
   std::string name{};
@@ -1754,7 +1818,7 @@ class BluetoothLERawAdvertisementsResponse : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 93;
   static constexpr uint16_t ESTIMATED_SIZE = 34;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "bluetooth_le_raw_advertisements_response"; }
+  const char *message_name() const override;
 #endif
   std::vector<BluetoothLERawAdvertisement> advertisements{};
   void encode(ProtoWriteBuffer buffer) const override;
@@ -1771,7 +1835,7 @@ class BluetoothDeviceRequest : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 68;
   static constexpr uint16_t ESTIMATED_SIZE = 12;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "bluetooth_device_request"; }
+  const char *message_name() const override;
 #endif
   uint64_t address{0};
   enums::BluetoothDeviceRequestType request_type{};
@@ -1791,7 +1855,7 @@ class BluetoothDeviceConnectionResponse : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 69;
   static constexpr uint16_t ESTIMATED_SIZE = 14;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "bluetooth_device_connection_response"; }
+  const char *message_name() const override;
 #endif
   uint64_t address{0};
   bool connected{false};
@@ -1811,7 +1875,7 @@ class BluetoothGATTGetServicesRequest : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 70;
   static constexpr uint16_t ESTIMATED_SIZE = 4;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "bluetooth_gatt_get_services_request"; }
+  const char *message_name() const override;
 #endif
   uint64_t address{0};
   void encode(ProtoWriteBuffer buffer) const override;
@@ -1823,6 +1887,7 @@ class BluetoothGATTGetServicesRequest : public ProtoMessage {
  protected:
   bool decode_field(uint32_t field_id, ProtoFieldValue value) override;
 };
+#endif
 class BluetoothGATTDescriptor : public ProtoMessage {
  public:
   std::vector<uint64_t> uuid{};
@@ -1851,6 +1916,7 @@ class BluetoothGATTCharacteristic : public ProtoMessage {
  protected:
   bool decode_field(uint32_t field_id, ProtoFieldValue value) override;
 };
+#ifdef USE_BLUETOOTH_PROXY
 class BluetoothGATTService : public ProtoMessage {
  public:
   std::vector<uint64_t> uuid{};
@@ -1870,7 +1936,7 @@ class BluetoothGATTGetServicesResponse : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 71;
   static constexpr uint16_t ESTIMATED_SIZE = 38;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "bluetooth_gatt_get_services_response"; }
+  const char *message_name() const override;
 #endif
   uint64_t address{0};
   std::vector<BluetoothGATTService> services{};
@@ -1888,7 +1954,7 @@ class BluetoothGATTGetServicesDoneResponse : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 72;
   static constexpr uint16_t ESTIMATED_SIZE = 4;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "bluetooth_gatt_get_services_done_response"; }
+  const char *message_name() const override;
 #endif
   uint64_t address{0};
   void encode(ProtoWriteBuffer buffer) const override;
@@ -1905,7 +1971,7 @@ class BluetoothGATTReadRequest : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 73;
   static constexpr uint16_t ESTIMATED_SIZE = 8;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "bluetooth_gatt_read_request"; }
+  const char *message_name() const override;
 #endif
   uint64_t address{0};
   uint32_t handle{0};
@@ -1923,7 +1989,7 @@ class BluetoothGATTReadResponse : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 74;
   static constexpr uint16_t ESTIMATED_SIZE = 17;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "bluetooth_gatt_read_response"; }
+  const char *message_name() const override;
 #endif
   uint64_t address{0};
   uint32_t handle{0};
@@ -1942,7 +2008,7 @@ class BluetoothGATTWriteRequest : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 75;
   static constexpr uint16_t ESTIMATED_SIZE = 19;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "bluetooth_gatt_write_request"; }
+  const char *message_name() const override;
 #endif
   uint64_t address{0};
   uint32_t handle{0};
@@ -1962,7 +2028,7 @@ class BluetoothGATTReadDescriptorRequest : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 76;
   static constexpr uint16_t ESTIMATED_SIZE = 8;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "bluetooth_gatt_read_descriptor_request"; }
+  const char *message_name() const override;
 #endif
   uint64_t address{0};
   uint32_t handle{0};
@@ -1980,7 +2046,7 @@ class BluetoothGATTWriteDescriptorRequest : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 77;
   static constexpr uint16_t ESTIMATED_SIZE = 17;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "bluetooth_gatt_write_descriptor_request"; }
+  const char *message_name() const override;
 #endif
   uint64_t address{0};
   uint32_t handle{0};
@@ -1999,7 +2065,7 @@ class BluetoothGATTNotifyRequest : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 78;
   static constexpr uint16_t ESTIMATED_SIZE = 10;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "bluetooth_gatt_notify_request"; }
+  const char *message_name() const override;
 #endif
   uint64_t address{0};
   uint32_t handle{0};
@@ -2018,7 +2084,7 @@ class BluetoothGATTNotifyDataResponse : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 79;
   static constexpr uint16_t ESTIMATED_SIZE = 17;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "bluetooth_gatt_notify_data_response"; }
+  const char *message_name() const override;
 #endif
   uint64_t address{0};
   uint32_t handle{0};
@@ -2037,7 +2103,7 @@ class SubscribeBluetoothConnectionsFreeRequest : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 80;
   static constexpr uint16_t ESTIMATED_SIZE = 0;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "subscribe_bluetooth_connections_free_request"; }
+  const char *message_name() const override;
 #endif
 #ifdef HAS_PROTO_MESSAGE_DUMP
   void dump_to(std::string &out) const override;
@@ -2050,7 +2116,7 @@ class BluetoothConnectionsFreeResponse : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 81;
   static constexpr uint16_t ESTIMATED_SIZE = 16;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "bluetooth_connections_free_response"; }
+  const char *message_name() const override;
 #endif
   uint32_t free{0};
   uint32_t limit{0};
@@ -2069,7 +2135,7 @@ class BluetoothGATTErrorResponse : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 82;
   static constexpr uint16_t ESTIMATED_SIZE = 12;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "bluetooth_gatt_error_response"; }
+  const char *message_name() const override;
 #endif
   uint64_t address{0};
   uint32_t handle{0};
@@ -2088,7 +2154,7 @@ class BluetoothGATTWriteResponse : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 83;
   static constexpr uint16_t ESTIMATED_SIZE = 8;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "bluetooth_gatt_write_response"; }
+  const char *message_name() const override;
 #endif
   uint64_t address{0};
   uint32_t handle{0};
@@ -2106,7 +2172,7 @@ class BluetoothGATTNotifyResponse : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 84;
   static constexpr uint16_t ESTIMATED_SIZE = 8;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "bluetooth_gatt_notify_response"; }
+  const char *message_name() const override;
 #endif
   uint64_t address{0};
   uint32_t handle{0};
@@ -2124,7 +2190,7 @@ class BluetoothDevicePairingResponse : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 85;
   static constexpr uint16_t ESTIMATED_SIZE = 10;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "bluetooth_device_pairing_response"; }
+  const char *message_name() const override;
 #endif
   uint64_t address{0};
   bool paired{false};
@@ -2143,7 +2209,7 @@ class BluetoothDeviceUnpairingResponse : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 86;
   static constexpr uint16_t ESTIMATED_SIZE = 10;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "bluetooth_device_unpairing_response"; }
+  const char *message_name() const override;
 #endif
   uint64_t address{0};
   bool success{false};
@@ -2162,7 +2228,7 @@ class UnsubscribeBluetoothLEAdvertisementsRequest : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 87;
   static constexpr uint16_t ESTIMATED_SIZE = 0;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "unsubscribe_bluetooth_le_advertisements_request"; }
+  const char *message_name() const override;
 #endif
 #ifdef HAS_PROTO_MESSAGE_DUMP
   void dump_to(std::string &out) const override;
@@ -2175,7 +2241,7 @@ class BluetoothDeviceClearCacheResponse : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 88;
   static constexpr uint16_t ESTIMATED_SIZE = 10;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "bluetooth_device_clear_cache_response"; }
+  const char *message_name() const override;
 #endif
   uint64_t address{0};
   bool success{false};
@@ -2194,7 +2260,7 @@ class BluetoothScannerStateResponse : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 126;
   static constexpr uint16_t ESTIMATED_SIZE = 4;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "bluetooth_scanner_state_response"; }
+  const char *message_name() const override;
 #endif
   enums::BluetoothScannerState state{};
   enums::BluetoothScannerMode mode{};
@@ -2212,7 +2278,7 @@ class BluetoothScannerSetModeRequest : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 127;
   static constexpr uint16_t ESTIMATED_SIZE = 2;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "bluetooth_scanner_set_mode_request"; }
+  const char *message_name() const override;
 #endif
   enums::BluetoothScannerMode mode{};
   void encode(ProtoWriteBuffer buffer) const override;
@@ -2224,12 +2290,14 @@ class BluetoothScannerSetModeRequest : public ProtoMessage {
  protected:
   bool decode_field(uint32_t field_id, ProtoFieldValue value) override;
 };
+#endif
+#ifdef USE_VOICE_ASSISTANT
 class SubscribeVoiceAssistantRequest : public ProtoMessage {
  public:
   static constexpr uint16_t MESSAGE_TYPE = 89;
   static constexpr uint16_t ESTIMATED_SIZE = 6;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "subscribe_voice_assistant_request"; }
+  const char *message_name() const override;
 #endif
   bool subscribe{false};
   uint32_t flags{0};
@@ -2261,7 +2329,7 @@ class VoiceAssistantRequest : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 90;
   static constexpr uint16_t ESTIMATED_SIZE = 41;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "voice_assistant_request"; }
+  const char *message_name() const override;
 #endif
   bool start{false};
   std::string conversation_id{};
@@ -2282,7 +2350,7 @@ class VoiceAssistantResponse : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 91;
   static constexpr uint16_t ESTIMATED_SIZE = 6;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "voice_assistant_response"; }
+  const char *message_name() const override;
 #endif
   uint32_t port{0};
   bool error{false};
@@ -2313,7 +2381,7 @@ class VoiceAssistantEventResponse : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 92;
   static constexpr uint16_t ESTIMATED_SIZE = 36;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "voice_assistant_event_response"; }
+  const char *message_name() const override;
 #endif
   enums::VoiceAssistantEvent event_type{};
   std::vector<VoiceAssistantEventData> data{};
@@ -2331,7 +2399,7 @@ class VoiceAssistantAudio : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 106;
   static constexpr uint16_t ESTIMATED_SIZE = 11;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "voice_assistant_audio"; }
+  const char *message_name() const override;
 #endif
   std::string data{};
   bool end{false};
@@ -2349,7 +2417,7 @@ class VoiceAssistantTimerEventResponse : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 115;
   static constexpr uint16_t ESTIMATED_SIZE = 30;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "voice_assistant_timer_event_response"; }
+  const char *message_name() const override;
 #endif
   enums::VoiceAssistantTimerEvent event_type{};
   std::string timer_id{};
@@ -2371,7 +2439,7 @@ class VoiceAssistantAnnounceRequest : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 119;
   static constexpr uint16_t ESTIMATED_SIZE = 29;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "voice_assistant_announce_request"; }
+  const char *message_name() const override;
 #endif
   std::string media_id{};
   std::string text{};
@@ -2391,7 +2459,7 @@ class VoiceAssistantAnnounceFinished : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 120;
   static constexpr uint16_t ESTIMATED_SIZE = 2;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "voice_assistant_announce_finished"; }
+  const char *message_name() const override;
 #endif
   bool success{false};
   void encode(ProtoWriteBuffer buffer) const override;
@@ -2422,7 +2490,7 @@ class VoiceAssistantConfigurationRequest : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 121;
   static constexpr uint16_t ESTIMATED_SIZE = 0;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "voice_assistant_configuration_request"; }
+  const char *message_name() const override;
 #endif
 #ifdef HAS_PROTO_MESSAGE_DUMP
   void dump_to(std::string &out) const override;
@@ -2435,7 +2503,7 @@ class VoiceAssistantConfigurationResponse : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 122;
   static constexpr uint16_t ESTIMATED_SIZE = 56;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "voice_assistant_configuration_response"; }
+  const char *message_name() const override;
 #endif
   std::vector<VoiceAssistantWakeWord> available_wake_words{};
   std::vector<std::string> active_wake_words{};
@@ -2454,7 +2522,7 @@ class VoiceAssistantSetConfiguration : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 123;
   static constexpr uint16_t ESTIMATED_SIZE = 18;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "voice_assistant_set_configuration"; }
+  const char *message_name() const override;
 #endif
   std::vector<std::string> active_wake_words{};
   void encode(ProtoWriteBuffer buffer) const override;
@@ -2466,12 +2534,14 @@ class VoiceAssistantSetConfiguration : public ProtoMessage {
  protected:
   bool decode_field(uint32_t field_id, ProtoFieldValue value) override;
 };
+#endif
+#ifdef USE_ALARM_CONTROL_PANEL
 class ListEntitiesAlarmControlPanelResponse : public InfoResponseProtoMessage {
  public:
   static constexpr uint16_t MESSAGE_TYPE = 94;
   static constexpr uint16_t ESTIMATED_SIZE = 57;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "list_entities_alarm_control_panel_response"; }
+  const char *message_name() const override;
 #endif
   uint32_t supported_features{0};
   bool requires_code{false};
@@ -2490,7 +2560,7 @@ class AlarmControlPanelStateResponse : public StateResponseProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 95;
   static constexpr uint16_t ESTIMATED_SIZE = 7;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "alarm_control_panel_state_response"; }
+  const char *message_name() const override;
 #endif
   enums::AlarmControlPanelState state{};
   void encode(ProtoWriteBuffer buffer) const override;
@@ -2507,7 +2577,7 @@ class AlarmControlPanelCommandRequest : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 96;
   static constexpr uint16_t ESTIMATED_SIZE = 16;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "alarm_control_panel_command_request"; }
+  const char *message_name() const override;
 #endif
   uint32_t key{0};
   enums::AlarmControlPanelStateCommand command{};
@@ -2521,12 +2591,14 @@ class AlarmControlPanelCommandRequest : public ProtoMessage {
  protected:
   bool decode_field(uint32_t field_id, ProtoFieldValue value) override;
 };
+#endif
+#ifdef USE_TEXT
 class ListEntitiesTextResponse : public InfoResponseProtoMessage {
  public:
   static constexpr uint16_t MESSAGE_TYPE = 97;
   static constexpr uint16_t ESTIMATED_SIZE = 68;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "list_entities_text_response"; }
+  const char *message_name() const override;
 #endif
   uint32_t min_length{0};
   uint32_t max_length{0};
@@ -2546,7 +2618,7 @@ class TextStateResponse : public StateResponseProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 98;
   static constexpr uint16_t ESTIMATED_SIZE = 16;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "text_state_response"; }
+  const char *message_name() const override;
 #endif
   std::string state{};
   bool missing_state{false};
@@ -2564,7 +2636,7 @@ class TextCommandRequest : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 99;
   static constexpr uint16_t ESTIMATED_SIZE = 14;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "text_command_request"; }
+  const char *message_name() const override;
 #endif
   uint32_t key{0};
   std::string state{};
@@ -2577,12 +2649,14 @@ class TextCommandRequest : public ProtoMessage {
  protected:
   bool decode_field(uint32_t field_id, ProtoFieldValue value) override;
 };
+#endif
+#ifdef USE_DATETIME_DATE
 class ListEntitiesDateResponse : public InfoResponseProtoMessage {
  public:
   static constexpr uint16_t MESSAGE_TYPE = 100;
   static constexpr uint16_t ESTIMATED_SIZE = 49;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "list_entities_date_response"; }
+  const char *message_name() const override;
 #endif
   void encode(ProtoWriteBuffer buffer) const override;
   void calculate_size(uint32_t &total_size) const override;
@@ -2598,7 +2672,7 @@ class DateStateResponse : public StateResponseProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 101;
   static constexpr uint16_t ESTIMATED_SIZE = 19;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "date_state_response"; }
+  const char *message_name() const override;
 #endif
   bool missing_state{false};
   uint32_t year{0};
@@ -2618,7 +2692,7 @@ class DateCommandRequest : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 102;
   static constexpr uint16_t ESTIMATED_SIZE = 17;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "date_command_request"; }
+  const char *message_name() const override;
 #endif
   uint32_t key{0};
   uint32_t year{0};
@@ -2633,12 +2707,14 @@ class DateCommandRequest : public ProtoMessage {
  protected:
   bool decode_field(uint32_t field_id, ProtoFieldValue value) override;
 };
+#endif
+#ifdef USE_DATETIME_TIME
 class ListEntitiesTimeResponse : public InfoResponseProtoMessage {
  public:
   static constexpr uint16_t MESSAGE_TYPE = 103;
   static constexpr uint16_t ESTIMATED_SIZE = 49;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "list_entities_time_response"; }
+  const char *message_name() const override;
 #endif
   void encode(ProtoWriteBuffer buffer) const override;
   void calculate_size(uint32_t &total_size) const override;
@@ -2654,7 +2730,7 @@ class TimeStateResponse : public StateResponseProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 104;
   static constexpr uint16_t ESTIMATED_SIZE = 19;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "time_state_response"; }
+  const char *message_name() const override;
 #endif
   bool missing_state{false};
   uint32_t hour{0};
@@ -2674,7 +2750,7 @@ class TimeCommandRequest : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 105;
   static constexpr uint16_t ESTIMATED_SIZE = 17;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "time_command_request"; }
+  const char *message_name() const override;
 #endif
   uint32_t key{0};
   uint32_t hour{0};
@@ -2689,12 +2765,14 @@ class TimeCommandRequest : public ProtoMessage {
  protected:
   bool decode_field(uint32_t field_id, ProtoFieldValue value) override;
 };
+#endif
+#ifdef USE_EVENT
 class ListEntitiesEventResponse : public InfoResponseProtoMessage {
  public:
   static constexpr uint16_t MESSAGE_TYPE = 107;
   static constexpr uint16_t ESTIMATED_SIZE = 76;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "list_entities_event_response"; }
+  const char *message_name() const override;
 #endif
   std::string device_class{};
   std::vector<std::string> event_types{};
@@ -2712,7 +2790,7 @@ class EventResponse : public StateResponseProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 108;
   static constexpr uint16_t ESTIMATED_SIZE = 14;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "event_response"; }
+  const char *message_name() const override;
 #endif
   std::string event_type{};
   void encode(ProtoWriteBuffer buffer) const override;
@@ -2724,12 +2802,14 @@ class EventResponse : public StateResponseProtoMessage {
  protected:
   bool decode_field(uint32_t field_id, ProtoFieldValue value) override;
 };
+#endif
+#ifdef USE_VALVE
 class ListEntitiesValveResponse : public InfoResponseProtoMessage {
  public:
   static constexpr uint16_t MESSAGE_TYPE = 109;
   static constexpr uint16_t ESTIMATED_SIZE = 64;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "list_entities_valve_response"; }
+  const char *message_name() const override;
 #endif
   std::string device_class{};
   bool assumed_state{false};
@@ -2749,7 +2829,7 @@ class ValveStateResponse : public StateResponseProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 110;
   static constexpr uint16_t ESTIMATED_SIZE = 12;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "valve_state_response"; }
+  const char *message_name() const override;
 #endif
   float position{0.0f};
   enums::ValveOperation current_operation{};
@@ -2767,7 +2847,7 @@ class ValveCommandRequest : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 111;
   static constexpr uint16_t ESTIMATED_SIZE = 14;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "valve_command_request"; }
+  const char *message_name() const override;
 #endif
   uint32_t key{0};
   bool has_position{false};
@@ -2782,12 +2862,14 @@ class ValveCommandRequest : public ProtoMessage {
  protected:
   bool decode_field(uint32_t field_id, ProtoFieldValue value) override;
 };
+#endif
+#ifdef USE_DATETIME_DATETIME
 class ListEntitiesDateTimeResponse : public InfoResponseProtoMessage {
  public:
   static constexpr uint16_t MESSAGE_TYPE = 112;
   static constexpr uint16_t ESTIMATED_SIZE = 49;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "list_entities_date_time_response"; }
+  const char *message_name() const override;
 #endif
   void encode(ProtoWriteBuffer buffer) const override;
   void calculate_size(uint32_t &total_size) const override;
@@ -2803,7 +2885,7 @@ class DateTimeStateResponse : public StateResponseProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 113;
   static constexpr uint16_t ESTIMATED_SIZE = 12;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "date_time_state_response"; }
+  const char *message_name() const override;
 #endif
   bool missing_state{false};
   uint32_t epoch_seconds{0};
@@ -2821,7 +2903,7 @@ class DateTimeCommandRequest : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 114;
   static constexpr uint16_t ESTIMATED_SIZE = 10;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "date_time_command_request"; }
+  const char *message_name() const override;
 #endif
   uint32_t key{0};
   uint32_t epoch_seconds{0};
@@ -2834,12 +2916,14 @@ class DateTimeCommandRequest : public ProtoMessage {
  protected:
   bool decode_field(uint32_t field_id, ProtoFieldValue value) override;
 };
+#endif
+#ifdef USE_UPDATE
 class ListEntitiesUpdateResponse : public InfoResponseProtoMessage {
  public:
   static constexpr uint16_t MESSAGE_TYPE = 116;
   static constexpr uint16_t ESTIMATED_SIZE = 58;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "list_entities_update_response"; }
+  const char *message_name() const override;
 #endif
   std::string device_class{};
   void encode(ProtoWriteBuffer buffer) const override;
@@ -2856,7 +2940,7 @@ class UpdateStateResponse : public StateResponseProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 117;
   static constexpr uint16_t ESTIMATED_SIZE = 61;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "update_state_response"; }
+  const char *message_name() const override;
 #endif
   bool missing_state{false};
   bool in_progress{false};
@@ -2881,7 +2965,7 @@ class UpdateCommandRequest : public ProtoMessage {
   static constexpr uint16_t MESSAGE_TYPE = 118;
   static constexpr uint16_t ESTIMATED_SIZE = 7;
 #ifdef HAS_PROTO_MESSAGE_DUMP
-  const char *message_name() const override { return "update_command_request"; }
+  const char *message_name() const override;
 #endif
   uint32_t key{0};
   enums::UpdateCommand command{};
@@ -2894,6 +2978,7 @@ class UpdateCommandRequest : public ProtoMessage {
  protected:
   bool decode_field(uint32_t field_id, ProtoFieldValue value) override;
 };
+#endif
 
 }  // namespace api
 }  // namespace esphome
